Fixed sound. Sampled in 8000Hz as .ogg files (.ogg does not work on iOS btw) and removed .wav versions.
Changed galloping sound to a Music object instead of Sound. Do not know why, but it could not be played at all as a Sound, and there was no trouble with playing it as a Music object.

Set volumes of the Sound and Music object to equalize difference between them.